Moved all FHIR operations into operations directory
