+++ packages/server/src/fhir/operations/expand.test.ts
@@ -1,12 +1,12 @@
 import { OperationOutcome } from '@medplum/fhirtypes';
 import express from 'express';
 import request from 'supertest';
-import { initApp } from '../app';
-import { loadTestConfig } from '../config';
-import { closeDatabase, initDatabase } from '../database';
-import { initTestAuth } from '../test.setup';
-import { initKeys } from '../oauth';
-import { seedDatabase } from '../seed';
+import { initApp } from '../../app';
+import { loadTestConfig } from '../../config';
+import { closeDatabase, initDatabase } from '../../database';
+import { initKeys } from '../../oauth';
+import { seedDatabase } from '../../seed';
+import { initTestAuth } from '../../test.setup';
 
 const app = express();
 let accessToken: string;
