+++ packages/server/src/fhir/operations/graphql.test.ts
@@ -5,13 +5,13 @@ import express from 'express';
 import { mkdtempSync, rmSync } from 'fs';
 import { sep } from 'path';
 import request from 'supertest';
-import { initApp } from '../app';
-import { loadTestConfig } from '../config';
-import { closeDatabase, initDatabase } from '../database';
-import { initKeys } from '../oauth';
-import { seedDatabase } from '../seed';
-import { initTestAuth } from '../test.setup';
-import { initBinaryStorage } from './storage';
+import { initApp } from '../../app';
+import { loadTestConfig } from '../../config';
+import { closeDatabase, initDatabase } from '../../database';
+import { initKeys } from '../../oauth';
+import { seedDatabase } from '../../seed';
+import { initTestAuth } from '../../test.setup';
+import { initBinaryStorage } from '../storage';
 
 const app = express();
 const binaryDir = mkdtempSync(__dirname + sep + 'binary-');
